requirements.txt
fastapi==0.115.6
python-multipart==0.0.9
pyyaml==6.0.3
requests==2.32.3
uvicorn[standard]==0.32.1
